Rework uploads to request macro
